requirements.txt
pillow == 10.2.0
rembg == 2.0.57
easygui == 0.98.3